Trying out ideas for control code.
The example code in baja.js is jsut copied from mozilla docs on gamepads.
I will have to try it with a real bluetooth controller.

The idea will be to set up a server that sends over telemetry, and receives control inputs for the servos and motor.

The initial manual control will allow me to iron out kinks in the controls, and prepare it for daisy.