Car Electronics Done!
At least I hope, cuz that battery compartment is jammed with stuff, and unding what I just did would be painful.

Everything looks like it is working. Only worry is that the Main Arm servo cable seemed to be wonky at some tests, but then I wasn't able to repeat it.

I still need to test the distance sensor to make sure it works.

At this point there are a total of 3 wildcards:
1) distance sensor
2) Steering Servo
3) mounting the arm - this might be the hardest