All components mounted!
Everything is working! I just need to mount the battery when it comes today, and the hardware is finished!

One weird thing that I came across is that the steering servo did weird stuff when I gave it different pule ranges - so I ended up limiting the actual servo output instead. It is still fine, it just requires explicit code to limit movement to a range other than 0-180