Hardware Finished
The robot is finally done! Everything seems to be working off of battery, including online-switch between battery and wall power. The online switch is finnicky, though, so it is important that it be done fast (if you push it without doing it hard enough, it loses contanct,but doesn't switch over to nextpower source in time.