Figured out audio... will take a while to fix
The issue was that the opus audio codec uses a packetization time of 20ms, which corresponds to 960 samples, and I was using 1024 or more.

Unfortunately, the packetization times are not arbitrary, and actually are required to be one of several specific values according to the webrtc standard.

This means that my code must add a translation layer, to reframe data that comes in with both larger and smaller frame sizes. I want to also do this in a way so that there is no audio dropout...

That just makes my life much harder.

The same thing can be said about video: max frame rate supported is 30fps. I need to make sure that the frame rate is capped in my code, and that I am giving aiortc legal packets - no matter what the subscribed stream is doing.

Wow, this is going to be tough...