Baja is now driving with video
However, there are several issues, so the library needs quite a bit more work to be robust to the Pi's failures.

Overall, with both audio and video streaming from the pi, it is running basically at maximum capacity, with all cores pegged at 50-60%.
This is not great, but not much can be done about that right now, other than waiting for pi 4, which should be more than powerful enough, if pi 3 can barely do it.

Nevertheless, if I can make it work well with the pi 3, it will be amazing.