Fixed cleanup code, separated js, and makefile
This is a general code cleanup. Code was working, but always had dirty exits. Now, exits are clean. This also fixed the dirty exit in testing.

Moved javascript over to its own area, this allows me to publish the module to npm, and to auto-minify it, and also to split the code up into files,
since it was getting pretty large.

Finally, moved all build steps to a makefile.

Lastly, the javascript wasn't being included in python packages. This fixed that issue also.